# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Test task","description":"Rest APIs"}],"paths":{"/files/uploads":{"get":{"tags":["Test task"],"summary":"Get filtered and sorted animals, example format:(http://localhost:8080/files/uploads?page=0&size=7&sort=id,desc&type=cat)","operationId":"getFilteredAndSortedAnimals","parameters":[{"name":"pageable","in":"query","description":"Pagination and sorting parameters","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}},{"name":"type","in":"query","description":"Type of animal","required":false,"schema":{"type":"string"}},{"name":"category","in":"query","description":"Category of animal","required":false,"schema":{"type":"string"}},{"name":"sex","in":"query","description":"Sex of animal","required":false,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageAnimalDto"}}}},"200":{"description":"Successfully retrieved filtered and sorted animals","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageAnimalDto"}}}}}},"post":{"tags":["Test task"],"summary":"Upload animals from file (CSV or XML)","operationId":"uploadAnimals","requestBody":{"content":{"multipart/form-data":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","description":"The file containing animals to upload","format":"binary"}}}}}},"responses":{"201":{"description":"Successfully uploaded animals","content":{"*/*":{"schema":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/AnimalDto"}}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/AnimalDto"}}}}}}}}},"components":{"schemas":{"AnimalDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"type":{"type":"string"},"sex":{"type":"string"},"weight":{"type":"integer","format":"int32"},"cost":{"type":"integer","format":"int32"},"category":{"type":"string"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"PageAnimalDto":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/AnimalDto"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"unpaged":{"type":"boolean"},"paged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"sorted":{"type":"boolean"},"empty":{"type":"boolean"},"unsorted":{"type":"boolean"}}}}}}